requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic[email]>=2.4.0
PyJWT>=2.8.0
passlib>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.6
